Fix bugs identified from kvmap comparison
- Fix fnv1a_hash32 using uint64_t instead of uint32_t for constants
- Fix type mismatches: use _kv_index for index parameters, size_t in loops
- Add validation requiring hash_func and key_cmp_func in kv_map_create
- Fix kv_map_print to check for NULL callbacks with pointer fallback
- Fix unused parameter warnings with (void) casts
- Use quoted includes for local headers
- Update tests to match new behavior

+++ fnv1a_hashes.c
@@ -1,4 +1,4 @@
-#include <fnv1a_hashes.h>
+#include "fnv1a_hashes.h"
 
 uint64_t fnv1a_hash64(const char* str) {
   static const uint64_t BASIS = 14695981039346656037ull;
@@ -12,8 +12,8 @@ uint64_t fnv1a_hash64(const char* str) {
 }
 
 uint32_t fnv1a_hash32(const char* str) {
-  static const uint64_t BASIS = 2166136261ul;
-  static const uint64_t FNV_PRIME = 16777619ull;
+  static const uint32_t BASIS = 2166136261ul;
+  static const uint32_t FNV_PRIME = 16777619ul;
   uint32_t hash = BASIS;
   for (const char* c = str; *c != '\0'; c++) {
     hash = hash ^ *c;
